moved getUserResult to utils directory
